+++ dbmodel/example/user/ud/ud_011_sample_other.sql
@@ -662,6 +662,8 @@ UPDATE cat_feature_node SET double_geom = '{"activated":true,"value":4}' WHERE i
 UPDATE cat_feature_node SET double_geom = '{"activated":true,"value":2}' WHERE id IN ('REGISTER');
 UPDATE cat_feature_node SET graph_delimiter='{DWFZONE}' WHERE id='WWTP';
 UPDATE cat_feature_node SET graph_delimiter='{DWFZONE}' WHERE id='OUTFALL';
+UPDATE cat_feature_node SET graph_delimiter='{OMUNIT}' WHERE id='CIRC_MANHOLE';
+UPDATE cat_feature_node SET graph_delimiter='{OMUNIT}' WHERE id='RECT_MANHOLE';
 UPDATE cat_feature_gully SET double_geom = '{"activated":true,"value":1}' WHERE id = 'GULLY';
 
 INSERT INTO config_param_user (parameter, value, cur_user) VALUES ('om_visit_status_vdefault', '4', current_user) ON CONFLICT (parameter, cur_user) DO NOTHING;