+++ dbmodel/test/ud/schema/views/test_schema_ve_node.sql
@@ -71,12 +71,12 @@ SELECT col_type_is('ve_node', 'macroexpl_id', 'int4', 'Column macroexpl_id shoul
 SELECT col_type_is('ve_node', 'muni_id', 'int4', 'Column muni_id should be int4');
 SELECT col_type_is('ve_node', 'sector_id', 'int4', 'Column sector_id should be int4');
 SELECT col_type_is('ve_node', 'macrosector_id', 'int4', 'Column macrosector_id should be int4');
-SELECT col_type_is('ve_node', 'sector_type', 'varchar(16)', 'Column sector_type should be varchar(16)');
+SELECT col_type_is('ve_node', 'sector_type', 'varchar(30)', 'Column sector_type should be varchar(30)');
 SELECT col_type_is('ve_node', 'drainzone_id', 'int4', 'Column drainzone_id should be int4');
-SELECT col_type_is('ve_node', 'drainzone_type', 'varchar(16)', 'Column drainzone_type should be varchar(16)');
+SELECT col_type_is('ve_node', 'drainzone_type', 'varchar(30)', 'Column drainzone_type should be varchar(30)');
 SELECT col_type_is('ve_node', 'drainzone_outfall', 'int4[]', 'Column drainzone_outfall should be int4[]');
 SELECT col_type_is('ve_node', 'dwfzone_id', 'int4', 'Column dwfzone_id should be int4');
-SELECT col_type_is('ve_node', 'dwfzone_type', 'varchar(16)', 'Column dwfzone_type should be varchar(16)');
+SELECT col_type_is('ve_node', 'dwfzone_type', 'varchar(30)', 'Column dwfzone_type should be varchar(30)');
 SELECT col_type_is('ve_node', 'dwfzone_outfall', 'int4[]', 'Column dwfzone_outfall should be int4[]');
 SELECT col_type_is('ve_node', 'omzone_id', 'int4', 'Column omzone_id should be int4');
 SELECT col_type_is('ve_node', 'macroomzone_id', 'int4', 'Column macroomzone_id should be int4');