+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Algorithm to calculate omunits and macroomunits `gw_fct_graphanalytics_omunit`.
+- `has_access` column to `node` table.
+- `ve_omunit` and `ve_macroomunit` views with edit triggers.
+
 ### Removed
 
 - `sector_id` and `muni_id` from `exploitation` table.